requirements.txt
Telethon==1.24.0
python-dotenv~=0.19.2
PyMySQL~=1.0.2
SQLAlchemy~=1.4.29
pytest~=6.2.5